Improve rosjava compatibility with Mac OS X
git-svn-id: file:///home/dthomas/wg/migration/svn2git/ros-svn/stacks/ros_experimental/trunk@7214 61973afe-1cd6-434e-a0a9-934cb0052259

+++ rosjava/rosjava.cpp
@@ -550,7 +550,7 @@ JNIEXPORT void JNICALL Java_ros_roscpp_JNI_setParam__JLjava_lang_String_2I
   (JNIEnv * env, jclass __jni, jlong cppHandle, jstring jparam, jint val)
 {
 	NodeHandle *handle = (NodeHandle *) cppHandle;
-	return handle->setParam(getString(env, jparam), val);
+	return handle->setParam(getString(env, jparam), (int)val);
 }
